Make a new .NET serialization lib
